Update access token

+++ data/src/main/java/com/jesusbadenas/goldenspearchallenge/data/api/response/ApiTokenResponse.kt
@@ -7,5 +7,5 @@ import com.google.gson.annotations.SerializedName
 data class ApiTokenResponse(
     @SerializedName("access_token") val accessToken: String,
     @SerializedName("token_type") val tokenType: String,
-    @SerializedName("expires_in") val expiresIn: Int
+    @SerializedName("expires_in") val expiresIn: Long
 )
